# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Hospital Management Solution",
      "version": "1.0.0",
      "title": "Hospital Management Solution API documentation",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "servers": [
      {
        "url": "/api/"
      }
    ],
...
    "paths": {
...
      "/doctor/add-doctor-info/": {
        "post": {
          "tags": [
            "Doctors Info"
          ],
          "summary": "Create a new doctor info record",
          "requestBody": {
            "description": "Create a new doctor info record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/doctorAddInfo"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": ""
            }
          }
        }
      },
      "/doctor/fetch-doctor-info/": {
        "get": {
          "tags": [
            "Doctors Info"
          ],
          "summary": "Fetch doctor info record of all doctors",
          "parameters": [
            {
                "name": "page_no",
                "in": "query",
                "description": "Page number",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 0
              },
              {
                "name": "page_size",
                "in": "query",
                "description": "Page size",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 1
              },
              {
                "name": "doctor_email",
                "in": "query",
                "description": "filter by email value",
                "required": true,
                "type": "string",
                "example": "user@example"
              }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/doctor/fetch-doctor-info/{doctor_id}": {
        "get": {
          "tags": [
            "Doctors Info"
          ],
          "summary": "Fetch doctor info record of a particular doctor by id",
          "parameters": [
            {
                "name": "doctor_id",
                "in": "path",
                "description": "ID of the doctor to fetch",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 2
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/doctor/modify-doctor-info/{doctor_id}": {
        "put": {
          "tags": [
            "Doctors Info"
          ],
          "summary": "Modify doctor information by ID",
          "parameters": [
            {
              "name": "doctor_id",
              "in": "path",
              "description": "ID of the doctor to modify",
              "required": true,
              "type": "integer",
              "format": "int32",
              "example": 2
            }
          ],
          "requestBody": {
            "description": "Edit a doctor info record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/doctorEditInfo"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "",
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "/doctor/add-doctor-availability/": {
        "post": {
          "tags": [
            "Doctors Availability Info"
          ],
          "summary": "Create a new doctor availability record",
          "requestBody": {
            "description": "Create a new doctor availability record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/doctorAddAvailability"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": ""
            }
          }
        }
      },
      "/doctor/fetch-doctor-availability/{doctor_id}": {
        "get": {
          "tags": [
            "Doctors Availability Info"
          ],
          "summary": "Fetch docotor availability info record of a particular doctor by id",
          "parameters": [
            {
                "name": "doctor_id",
                "in": "path",
                "description": "ID of the doctor to fetch",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 2
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/doctor/modify-doctor-availability/{doctor_id}": {
        "put": {
          "tags": [
            "Doctors Availability Info"
          ],
          "summary": "Modify doctor availability information by ID",
          "parameters": [
            {
              "name": "doctor_id",
              "in": "path",
              "description": "ID of the doctor to modify",
              "required": true,
              "type": "integer",
              "format": "int32",
              "example": 2
            }
          ],
          "requestBody": {
            "description": "Edit a doctor availability info record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/doctorEditAvailability"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "",
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "/doctor/fetch-assigned-patients/{doctor_id}": {
        "get": {
          "tags": [
            "Doctor Patient Assigned"
          ],
          "summary": "Fetch patients assigned to a particular doctor by id",
          "parameters": [
            {
                "name": "doctor_id",
                "in": "path",
                "description": "ID of the doctor to fetch",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 2
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/department/add-department-info/": {
        "post": {
          "tags": [
            "Department Info"
          ],
          "summary": "Create a new department info record",
          "requestBody": {
            "description": "Create a new department info record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/departmentAddInfo"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": ""
            }
          }
        }
      },
      "/department/fetch-department-info/": {
        "get": {
          "tags": [
            "Department Info"
          ],
          "summary": "Fetch department info record of all department",
          "parameters": [
            {
                "name": "page_no",
                "in": "query",
                "description": "Page number",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 0
              },
              {
                "name": "page_size",
                "in": "query",
                "description": "Page size",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 1
              },
              {
                "name": "department_name",
                "in": "query",
                "description": "filter by department name value",
                "required": true,
                "type": "string",
                "example": "Radiology"
              }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/department/fetch-department-info/{department_id}": {
        "get": {
          "tags": [
            "Department Info"
          ],
          "summary": "Fetch department info record of a particular department by id",
          "parameters": [
            {
                "name": "department_id",
                "in": "path",
                "description": "ID of the department to fetch",
                "required": true,
                "type": "integer",
                "format": "int32",
                "example": 2
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": ""
            }
          }
        }
      },
      "/department/modify-department-info/{department_id}": {
        "put": {
          "tags": [
            "Department Info"
          ],
          "summary": "Modify department information by ID",
          "parameters": [
            {
              "name": "department_id",
              "in": "path",
              "description": "ID of the department to modify",
              "required": true,
              "type": "integer",
              "format": "int32",
              "example": 2
            }
          ],
          "requestBody": {
            "description": "Edit a department info record",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/departmentEditInfo"
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "",
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "doctorAddInfo": {
            "type": "object",
            "properties": {
              "department_id": {
                  "type": "number",
                  "format": "number",
                  "example": 1
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "name": {
                "type": "string",
                "format": "string",
                "example": "John Doe"
              },
              "phone": {
                "type": "string",
                "format": "string",
                "example": "+91 12345678"
              },
              "specialization": {
                "type": "string",
                "format": "string",
                "example": "Cardiologist"
              }
            }
        },
        "doctorEditInfo": {
            "type": "object",
            "properties": {
                "department_id": {
                    "type": "number",
                    "format": "number",
                    "example": 1
                },
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "name": {
                  "type": "string",
                  "format": "string",
                  "example": "John Doe"
                },
                "phone": {
                  "type": "string",
                  "format": "string",
                  "example": "+91 12345678"
                },
                "specialization": {
                  "type": "string",
                  "format": "string",
                  "example": "Pulmonologist"
                }
              }
        },
        "doctorAddAvailability": {
            "type": "object",
            "properties": {
              "doctor_id": {
                  "type": "number",
                  "format": "number",
                  "example": 1
              },
              "unavailabile_times": {
                "type": "array",
                "items": {
                "type": "string"
                },
                "example": ["1711818058", "1711869749"]
              },
              "available_times": {
                "type": "string",
                "format":"string",
                "example": [
                      {"1": ["1000", "1700"]},
                      {"2": ["1000", "1700"]},
                      {"3": ["1000", "1700"]},
                      {"4": ["1000", "1700"]},
                      {"5": ["1000", "1500"]}
                    ]                  
              }
            }
        },
        "doctorEditAvailability": {
            "type": "object",
            "properties": {
              "doctor_id": {
                  "type": "number",
                  "format": "number",
                  "example": 2
              }
            }
        },
        "departmentAddInfo": {
            "type": "object",
            "properties": {
              "department_name": {
                "type": "string",
                "format": "string",
                "example": "Radiology"
              },
              "services":{
                "type": "array",
                "items": {
                "type": "string"
                },
                "example": ["MRI", "CT SCAN"]
              }
            }
        },
        "departmentEditInfo": {
            "type": "object",
            "properties": {
              "services":{
                "type": "array",
                "items": {
                "type": "string"
                },
                "example": ["MRI", "CT SCAN", "UltraSound"]
              }
            }
        }
      }
    }
  }